// dashboard/ui/src/api/generated/models/experimentSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */
import type { ExperimentSourceType } from './experimentSourceType.ts';
import type { FileInfo } from './fileInfo.ts';

export interface ExperimentSource {
  type: ExperimentSourceType;
  pdb_id?: string;
  url?: string;
  files: FileInfo[];
  /** @minimum 0 */
  file_count?: number;
}